feat: add EVM Provider Context Ext
