+++ macros/context-derive/Cargo.toml
@@ -23,6 +23,10 @@ proc-macro2 = { workspace = true }
 [dev-dependencies]
 trybuild = { workspace = true }
 gadget-sdk = { workspace = true, features = ["std"] }
+# EVM Stuff
+alloy-network = { workspace = true }
+alloy-provider = { workspace = true }
+alloy-transport = { workspace = true }
 
 [features]
 default = ["std"]
