+++ macros/context-derive/tests/ui/01_basic.rs
@@ -1,7 +1,7 @@
 use gadget_sdk::config::StdGadgetConfiguration;
-use gadget_sdk::ctx::KeystoreContext;
+use gadget_sdk::ctx::{EVMProviderContext, KeystoreContext};
 
-#[derive(KeystoreContext)]
+#[derive(KeystoreContext, EVMProviderContext)]
 struct MyContext {
     foo: String,
     #[config]
@@ -14,4 +14,5 @@ fn main() {
         sdk_config: Default::default(),
     };
     let _keystore = ctx.keystore();
+    let _evm_provider = ctx.evm_provider();
 }
