+++ macros/context-derive/tests/ui/03_generic_struct.rs
@@ -1,7 +1,7 @@
 use gadget_sdk::config::StdGadgetConfiguration;
-use gadget_sdk::ctx::KeystoreContext;
+use gadget_sdk::ctx::{EVMProviderContext, KeystoreContext};
 
-#[derive(KeystoreContext)]
+#[derive(KeystoreContext, EVMProviderContext)]
 struct MyContext<T, U> {
     foo: T,
     bar: U,
@@ -16,4 +16,5 @@ fn main() {
         sdk_config: Default::default(),
     };
     let _keystore = ctx.keystore();
+    let _evm_provider = ctx.evm_provider();
 }
